requirements.txt
Flask~=2.2.2
azure-cognitiveservices-vision-computervision